Refactor DAO
